Removing name sorting from the RPC Interface config creation
Fixing a bug where RPC functions do not call their correct counterparts in certain scenarios, highlighted below:

When two separate projects use different code files for their RPC handling (in my case a project where the client and server project are seperated), sorting the names causes large issues. In my usecase, it was that if functions were not ordered the same way in both client and server projects, causing the wrong function to be called when a network message comes in.

What I would prefer to have as a fix is to synchronize the rpc config when connecting to the autorative, and detecting any mismatched rpc functions through communicating the strings from the server to client and back.